Refactor ruff code
